# Faster iteration