--- a/.travis.yml
+++ b/.travis.yml
@@ -8,13 +8,12 @@ env:
   - FEATURES='unsafe_internals'
 
 before_script:
-  - rustup component add rustfmt-preview
+  - "[ \"$TRAVIS_RUST_VERSION\" == 'nightly' ] || rustup component add rustfmt-preview"
 
 script:
-  - cargo fmt --all -- --check
   - cargo build --tests --features "$FEATURES"
   - cargo test --features "$FEATURES"
-  - "[ \"$TRAVIS_RUST_VERSION\" != 'nightly' ] ||  cargo bench --features \"$FEATURES\""
-  - "[ \"$TRAVIS_RUST_VERSION\" != 'nightly' ] ||  cargo fmt -- --check"
+  - "[ \"$TRAVIS_RUST_VERSION\" != 'nightly' ] || cargo bench --features \"$FEATURES\""
+  - "[ \"$TRAVIS_RUST_VERSION\" == 'nightly' ] || cargo fmt --all -- --check"
 
 cache: cargo