--- a/benches/vob.rs
+++ b/benches/vob.rs
@@ -118,3 +118,15 @@ fn iter_set_bits(bench: &mut Bencher) {
     a.extend((0..N).map(|_| rng.gen::<bool>()));
     bench.iter(|| a.iter_set_bits(..).count());
 }
+
+#[bench]
+fn iter_all_set_bits(bench: &mut Bencher) {
+    let a = Vob::from_elem(N, true);
+    bench.iter(|| a.iter_set_bits(..).count());
+}
+
+#[bench]
+fn iter_all_unset_bits(bench: &mut Bencher) {
+    let a = Vob::from_elem(N, true);
+    bench.iter(|| a.iter_unset_bits(..).count());
+}