@ltratt 2018-12-10
Annoyingly the tests are failing probably only because there isn't a `rustfmt` for nightly yet...