@ltratt 2018-12-10
Even if that latest commit does the trick, there's still going to be some squashing necessary...