@thomwiggers 2018-12-11
Right, I hadn't fully thought the probabilities through... But I guess this PR is strongly optimising for the everything is `T::max_value()` or `0` cases. Cases where those cases are interleaved in some weird way may be slightly different, but I'm a) not sure how realistic that is and b) too lazy to implement those benchmarks :stuck_out_tongue: