@thomwiggers 2018-12-11
You could squash, as far as I'm concerned. Perhaps re-order the commits a bit as well.