@ltratt 2018-12-11
Reordered and squashed. Note that I changed `CHANGES.md` to have the new release be today (rather than yesterday, which make it out of sync with `crates.io`).